move bulk of code from main_loop.c to chaos.c/h
