+++ src/chaos.h
@@ -1,0 +1,6 @@
+#ifndef PAPERMARIO_CHAOS_EDITION_CHAOS_H
+#define PAPERMARIO_CHAOS_EDITION_CHAOS_H
+
+void update_chaos(void);
+
+#endif // PAPERMARIO_CHAOS_EDITION_CHAOS_H
